Initial implementation of `AllowRestrictedCustomActions` statement.
This works, but we need to hack through TF 0.11's inability to create a nested
configuration block conditionally.  Ideally, the policy would not include a
the `AllowRestrictedCustomActions` at all if no custom actions were specified.
However, there is no way to express that directly in TF 0.11 syntax.

Keep in mind that statements must refer to valid S3 actions, so cannot be
something from `ec2:*` or `doesnotexist:*`.

As a workaround, the `AllowRestrictedCustomActions` statement defaults to what
is probably an innocuous action: `s3:GetAnalyticsConfiguration`.  The statement
does not allow any principals access by default, so no access should be granted
by the default configuration.